Restructure main
