+++ src/main.rs
@@ -1,73 +1,14 @@
 use actix_web::{web, App, HttpServer, Responder, HttpRequest, HttpResponse};
-use urlencoding::encode;
 
-use crate::utils::Message;
 extern crate chrono;
-use chrono::{DateTime, Utc};
+use chrono::Utc;
 
 mod utils;
+mod messages;
+use crate::messages::{search_serper_google, open_ai_response, return_open_ai_response};
 
 // HTTPS requests are getting forwarded from nginx to localhost
 const URL: &str = "127.0.0.1:61347";
-
-fn prepend_message(body: &mut web::Json<utils::OpenAIRequest>, message: &String){
-    let message  = utils::Message::new("system", &message);
-    body.messages.insert(0, message);
-}
-
-fn append_message(body: &mut web::Json<utils::OpenAIRequest>, message: &String){
-    let message  = utils::Message::new("system", &message);
-    body.messages.push(message);
-}
-
-async fn completions(req: HttpRequest, mut body: web::Json<utils::OpenAIRequest>) -> impl Responder {
-    // Check the API key in the HTTP header
-    if !utils::authorize(req){
-        return HttpResponse::Unauthorized().body("Invalid or missing API key")
-    }
-
-    // Load prompts from config
-    let config = utils::Config::new(String::from("config.json"));
-
-    let user_question = body.messages[0].clone();
-
-    let pose_second_query = body.model == "gpt-4-turbo";
-
-    // Create instructions and run first OpenAI request. 
-    prepend_message(&mut body, &config.first_instruction);
-    prepend_message(&mut body, &format!("The current time is {}",  Utc::now()));
-    prepend_message(&mut body, &config.examples);
-
-    let response = utils::open_ai_response(&*body).await;
-    
-    // Run first search query based on Models output and append results.
-    append_message(&mut body, &response.choices[0].message.content);
-    let query = &response.choices[0].message.content;
-    println!("Query 1: {}", query);
-    let search_results = utils::search_serper_google(query).await;
-    append_message(&mut body, &search_results);
-
-    // With GPT-4, run a second google search
-    if pose_second_query {
-        append_message(&mut body, &config.second_instruction);
-        let response = utils::open_ai_response(&*body).await;
-        append_message(&mut body, &response.choices[0].message.content);
-
-        let query = &response.choices[0].message.content;
-        println!("Query 2: {}", query);
-        let search_results = utils::search_serper_google(&query).await;
-        append_message(&mut body, &search_results)
-    }
-
-    // Insert final instructions
-    append_message(&mut body, &config.final_instruction);
-
-    // Repeat user question
-    body.messages.push(user_question);
- 
-    // Send modified request to OpenAI and return result
-    utils::return_open_ai_response(&*body).await
-}
 
 #[actix_web::main]
 async fn main() -> std::io::Result<()> {
@@ -78,4 +19,62 @@ async fn main() -> std::io::Result<()> {
     .bind(URL)?
     .run()
     .await
+}
+
+async fn completions(req: HttpRequest, mut body: web::Json<messages::OpenAIRequest>) -> impl Responder {
+    // Check the API key in the HTTP header
+    if !utils::authorize(req){
+        return HttpResponse::Unauthorized().body("Invalid or missing API key")
+    }
+
+    // Load prompts from config
+    let config = utils::Config::new(String::from("config.json"));
+
+    let user_question = body.messages[0].clone();
+
+    // Create instructions and ask model for Google query 
+    prepend_message(&mut body, &config.first_instruction);
+    prepend_message(&mut body, &format!("The current time is {}",  Utc::now()));
+    prepend_message(&mut body, &config.examples);
+
+    let response = open_ai_response(&*body).await;
+    
+    // Run first search query based on Models output and append results.
+    append_message(&mut body, &response.choices[0].message.content);
+    let query = &response.choices[0].message.content;
+    println!("Query 1: {}", query);
+    let search_results = search_serper_google(query).await;
+    append_message(&mut body, &search_results);
+
+    // With GPT-4, run a second google search
+    if body.model == "gpt-4-turbo" {
+        append_message(&mut body, &config.second_instruction);
+        let response = open_ai_response(&*body).await;
+        append_message(&mut body, &response.choices[0].message.content);
+
+        let query = &response.choices[0].message.content;
+        println!("Query 2: {}", query);
+        let search_results = search_serper_google(&query).await;
+        append_message(&mut body, &search_results)
+    }
+
+    // Insert final instructions
+    append_message(&mut body, &config.final_instruction);
+
+    // Repeat user question
+    body.messages.push(user_question);
+ 
+    // Send modified request to OpenAI and return result to user
+    return_open_ai_response(&*body).await
+}
+
+
+fn prepend_message(body: &mut web::Json<messages::OpenAIRequest>, message: &String){
+    let message  = messages::Message::new("system", &message);
+    body.messages.insert(0, message);
+}
+
+fn append_message(body: &mut web::Json<messages::OpenAIRequest>, message: &String){
+    let message  = messages::Message::new("system", &message);
+    body.messages.push(message);
 }
